Create sonoff-S55.yaml
Specially modified version that allows the DALAS thermometer to be powered instead of the original button.
It is necessary to find the address of the thermometer in the log and then enter it into the paremeters. If you don't need the thermometer, just replace the entity with the button. That part is just commented out.
I made it so it could be used as a simple thermostat.
Of course you can use another thermometer with a hygrometer. 
The power de takes from the flash pads and the data cable is soldered to the button lead.